fix: リファクタリング途中. iOS16.2でビルド成功を確認した
