+++ pubspec.yaml
@@ -34,7 +34,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  google_mobile_ads: ^0.13.5
+  google_mobile_ads: ^3.0.0
   
 
 dev_dependencies:
